requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
spotipy>=2.23.0
nltk>=3.8.1
textblob>=0.17.1
scikit-learn>=1.3.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
python-dateutil>=2.8.2